requirements.txt
discord.py~=1.5.1
python-dotenv~=0.15.0
requests~=2.24.0
beautifulsoup4~=4.9.1